revamp CLI interface
Change the CLI interface to make it suitable for running go generate.
By default we generate just a single file containing the API
interface and the parameters that it uses.

When running a server, by default we generate the code to a temporary
directory and remove it afterwards.

We change the --port option to --http (mirroring the godoc flag)
so that it's possible to choose an address to listen on.